* Fixed a couple of Windows API calls where warning was generated for size_t being passed to DWORD argument generated warnings on x64 build. * Fixed a few cascading size_t to int conversion warnings on x64 build.

+++ boost/cgi/connections/anonymous_pipe.hpp
@@ -105,7 +105,7 @@ BOOST_CGI_NAMESPACE_BEGIN
       DWORD bytesRead;
       int ret = -1;
 
-      if (::ReadFile(file_handle, buf, len, &bytesRead, NULL))
+      if (::ReadFile(file_handle, buf, DWORD(len), &bytesRead, NULL))
         ret = bytesRead;
       else
         ec = boost::system::error_code(
@@ -142,7 +142,7 @@ BOOST_CGI_NAMESPACE_BEGIN
       DWORD bytesWritten;
       int ret = -1;
 
-      if (::WriteFile(file_handle, buf, len, &bytesWritten, NULL))
+      if (::WriteFile(file_handle, buf, DWORD(len), &bytesWritten, NULL))
         ret = bytesWritten;
       else
         ec = boost::system::error_code(
